Add initial project files and Employee class
In this commit, we initialize the TinkeringApp project. This includes the addition of configuration and project files, such as global.json and TinkeringApp.sln. A .gitignore is also added to avoid tracking unnecessary files.

Specifically, a new class "ValidQId" is introduced to validate the QId string inserted via the console input. This helps ensure that only numerical string values are valid for the QId property of the new Employee class. This is needed to ensure data consistency and integrity in the application.