Add Bogus library and implement employee data generation
Bogus library has been added to the project for generating mock employee data. The code now features the usage of Bogus to create a fake instance of an Employee, which is then mapped to an EmployeeDTO. This is followed by printing the DTO as a JSON object. An Employee model, along with its map method, has been also introduced.