Add EmployeeMapper for Employee and EmployeeDto conversion
The changes introduce an EmployeeMapper static class responsible for mapping between the Employee and EmployeeDto object types. The mapping responsibility has been moved out of the Employee model to enhance separation of concerns. The EmployeeDto has been modified from a record to a class. Also, the Program.cs file has been adjusted to utilize the new mapper and generate the EmployeeDto object accordingly.